Changed DatasetTable.onRowClick to emitter
