+++ src/app/datasets/anonymous-dashboard/anonymous-dashboard.component.html
@@ -8,6 +8,7 @@
         <dataset-table
           [tableColumns]="tableColumns$ | async"
           (settingsClick)="onSettingsClick()"
+          (rowClick)="onRowClick($event)"
         ></dataset-table>
       </div>
     </div>
