+++ src/app/datasets/dashboard/dashboard.component.spec.ts
@@ -56,6 +56,9 @@ describe("DashboardComponent", () => {
   let component: DashboardComponent;
   let fixture: ComponentFixture<DashboardComponent>;
 
+  const router = {
+    navigateByUrl: jasmine.createSpy("navigateByUrl")
+  };
   let store: MockStore;
   let dispatchSpy;
 
@@ -83,7 +86,7 @@ describe("DashboardComponent", () => {
           { provide: APP_CONFIG, useValue: { shoppingCartOnHeader: "true" } },
           { provide: ActivatedRoute, useClass: MockActivatedRoute },
           { provide: MatDialog, useClass: MockMatDialog },
-          { provide: Router, useClass: MockRouter }
+          { provide: Router, useValue: router }
         ]
       }
     });
@@ -188,6 +191,18 @@ describe("DashboardComponent", () => {
       expect(dispatchSpy).toHaveBeenCalledTimes(1);
       expect(dispatchSpy).toHaveBeenCalledWith(
         deselectColumnAction({ name: column.name, columnType: column.type })
+      );
+    });
+  });
+
+  describe("#onRowClick()", () => {
+    it("should navigate to a dataset", () => {
+      const dataset = new Dataset();
+      component.onRowClick(dataset);
+
+      expect(router.navigateByUrl).toHaveBeenCalledTimes(1);
+      expect(router.navigateByUrl).toHaveBeenCalledWith(
+        "/datasets/" + encodeURIComponent(dataset.pid)
       );
     });
   });
