+++ src/app/datasets/dataset-table/dataset-table.component.html
@@ -375,7 +375,7 @@
     <mat-row
       class="row-generic"
       *matRowDef="let dataset; columns: displayedColumns"
-      (click)="onRowClick(dataset)"
+      (click)="doRowClick(dataset)"
     >
     </mat-row>
   </mat-table>
